Fixed bug: Console input freezes on tab-complete when empty
Currently if you press tab when in console and have nothing written in your command line spigot/bukkit loops through EVERY command on the server and tries to find which one if them it is. This checks if the user has written something in the commandline or not. If not, do not perform tab-complete.