add setCanbePushed and isPushable to LivingEntity
Currently if we need to stop an entity from being pushed we can put a invisible armor stand on their head, however, this creates other problems that we must deal with. This change would allow developers to toggle whether or not a entity can be pushed by other entities. This would help keep entities stationary without using NMS.